Move code around to avoid circular imports.
